+++ src/Ext/AnimType/Body.h
@@ -6,9 +6,6 @@
 #include <Utilities/Enum.h>
 #include <Utilities/Constructs.h>
 #include <Utilities/Template.h>
-
-class AnimClass;
-
 class AnimTypeExt
 {
 public:
@@ -18,9 +15,24 @@ public:
 	{
 	public:
 		CustomPalette Palette;
+		Valueable<UnitTypeClass*> CreateUnit;
+		Valueable<unsigned short> CreateUnit_Facing;
+		Valueable<bool> CreateUnit_InheritDeathFacings;
+		Valueable<bool> CreateUnit_InheritTurretFacings;
+		Valueable<bool> CreateUnit_RemapAnim;
+		Valueable<bool> CreateUnit_RandomFacing;
+		Valueable<Mission> CreateUnit_Mission;
+		Valueable<OwnerHouseKind> CreateUnit_Owner;
 
-		ExtData(AnimTypeClass* OwnerObject) : Extension<AnimTypeClass>(OwnerObject),
-			Palette(CustomPalette::PaletteMode::Temperate)
+		ExtData(AnimTypeClass* OwnerObject) : Extension<AnimTypeClass>(OwnerObject)
+			, Palette(CustomPalette::PaletteMode::Temperate)
+			, CreateUnit_Facing(0)
+			, CreateUnit_RandomFacing(true)
+			, CreateUnit_InheritDeathFacings(false)
+			, CreateUnit_InheritTurretFacings(false)
+			, CreateUnit_RemapAnim(false)
+			, CreateUnit_Mission(Mission::Guard)
+			, CreateUnit_Owner(OwnerHouseKind::Victim)
 		{ }
 
 		virtual ~ExtData() = default;
@@ -45,4 +57,6 @@ public:
 	};
 
 	static ExtContainer ExtMap;
+
+	static const void ProcessDestroyAnims(UnitClass* pThis, TechnoClass* pKiller = nullptr);
 };