+++ src/Ext/Techno/Body.h
@@ -6,6 +6,7 @@
 #include <Utilities/TemplateDef.h>
 
 #include <New/Entity/ShieldClass.h>
+#include <New/Entity/LaserTrailClass.h>
 
 class BulletClass;
 
@@ -19,6 +20,9 @@ public:
 	public:
 		Valueable<BulletClass*> InterceptedBullet;
 		std::unique_ptr<ShieldClass> Shield;
+		Valueable<bool> WasCloaked;
+		ValueableVector<std::unique_ptr<LaserTrailClass>> LaserTrails;
+		Valueable<bool> ReceiveDamage;
 
 		Valueable<bool> WasCloaked;
         Valueable<TechnoClass*> KilledBy;
@@ -27,6 +31,8 @@ public:
 			InterceptedBullet(nullptr),
 			Shield(),
 			WasCloaked(false),
+			LaserTrails(),
+			ReceiveDamage(false)
             KilledBy(nullptr),
             LastKillWasTeamTarget(false)
 		{ }
@@ -62,6 +68,9 @@ public:
 
 	static bool IsHarvesting(TechnoClass* pThis);
 	static bool HasAvailableDock(TechnoClass* pThis);
+
+	static void InitializeLaserTrails(TechnoClass* pThis);
+	static CoordStruct GetFLHAbsoluteCoords(TechnoClass* pThis, CoordStruct flh, bool turretFLH = false);
 
 	static void TransferMindControlOnDeploy(TechnoClass* pTechnoFrom, TechnoClass* pTechnoTo);