+++ src/Phobos.Ext.cpp
@@ -2,6 +2,7 @@
 
 #include <Ext/Aircraft/Body.h>
 #include <Ext/AnimType/Body.h>
+#include <Ext/Anim/Body.h>
 #include <Ext/Building/Body.h>
 #include <Ext/BuildingType/Body.h>
 #include <Ext/Bullet/Body.h>
@@ -22,6 +23,7 @@
 #include <Ext/Team/Body.h>
 
 #include <New/Type/RadTypeClass.h>
+#include <New/Type/LaserTrailTypeClass.h>
 
 #include <utility>
 
@@ -221,6 +223,7 @@ auto MassActions = MassAction <
 	// Ext classes
 	AircraftExt,
 	AnimTypeExt,
+	AnimExt,
 	BuildingExt,
 	BuildingTypeExt,
 	BulletExt,
@@ -240,6 +243,7 @@ auto MassActions = MassAction <
 	WeaponTypeExt,
 	// New classes
 	ShieldTypeClass,
+	LaserTrailTypeClass,
 	RadTypeClass,
 	TeamExt
 	// other classes